Isolate GetDailyVerse(), GetRandomVerse(), and GetTrulyRandomVerse() into its own provider.
